@@ -3,25 +3,36 @@ package auth
 
 import (
 	"context"
-	"encoding/hex"
 	"encoding/json"
+	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"sort"
 	"strings"
 	"testing"
+	"time"
 
-	"github.com/gbrlsnchs/jwt/v3"
-	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/filecoin-project/venus-auth/config"
-	"github.com/filecoin-project/venus-auth/core"
-	"github.com/filecoin-project/venus-auth/storage"
+	"github.com/filecoin-project/go-address"
+
+	"github.com/ipfs-force-community/sophon-auth/config"
+	"github.com/ipfs-force-community/sophon-auth/core"
+	"github.com/ipfs-force-community/sophon-auth/storage"
 )
+
+var (
+	adminCtx = core.CtxWithPerm(context.Background(), core.PermAdmin)
+	signCtx  = core.CtxWithPerm(context.Background(), core.PermSign)
+	readCtx  = core.CtxWithPerm(context.Background(), core.PermRead)
+)
+
+func signPermAndRandomUsername() context.Context {
+	ctx := core.CtxWithPerm(context.Background(), core.PermSign)
+	return core.CtxWithName(ctx, time.Now().String())
+}
 
 func TestJwt(t *testing.T) {
 	limitStrs := `[{"Id":"794fc9a4-2b80-4503-835a-7e8e27360b3d","Name":"test_user_01","Service":"","API":"","ReqLimit":{"Cap":10,"ResetDur":120000000000}},{"Id":"252f581e-cbd2-4a61-a517-0b7df65013aa","Name":"test_user_02","Service":"","API":"","ReqLimit":{"Cap":10,"ResetDur":72000000000000}}]`
@@ -68,7 +79,7 @@ func TestJwt(t *testing.T) {
 	// stm: @VENUSAUTH_JWT_LIST_MINERS_001
 	t.Run("test list miner", func(t *testing.T) { testListMiner(t, userMiners) })
 	// stm: @VENUSAUTH_JWT_HAS_MINER_001, @VENUSAUTH_JWT_HAS_MINER_002
-	t.Run("test has miner", func(t *testing.T) { testHasMiner(t, userMiners) })
+	t.Run("test miner exist user", func(t *testing.T) { testMinerExistInMiner(t, userMiners) })
 	// stm: @VENUSAUTH_JWT_GET_USER_BY_MINER_001, @VENUSAUTH_JWT_GET_USER_BY_MINER_002, @VENUSAUTH_JWT_GET_USER_BY_MINER_003
 	t.Run("test get user by miner", func(t *testing.T) { testGetUserByMiner(t, userMiners) })
 	// stm: @VENUSAUTH_JWT_DELETE_MINER_001, @VENUSAUTH_JWT_DELETE_MINER_002
@@ -79,6 +90,7 @@ func TestJwt(t *testing.T) {
 		"test_user_001": {"t3wylwd6pclppme4qmbgwled5xpsbgwgqbn2alxa7yahg2gnbfkipsdv6m764xm5coizujmwdmkxeugplmorha", "t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua"},
 		"test_user_002": {"t3r47fkdzfmtex5ic3jnwlzc7bkpbj7s4d6limyt4f57t3cuqq5nuvhvwv2cu2a6iga2s64vjqcxjqiezyjooq", "t1uqtvvwkkfkkez52ocnqe6vg74qewiwja4t2tiba"},
 		"test_user_003": {"t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua", "t1sgeoaugenqnzftqp7wvwqebcozkxa5y7i56sy2q"},
+		"test_user_004": {"f410ffhviosf3a4ymvgm3sxrugj2qkir33qnlhqjrauq"},
 	}
 	t.Run("test register signer", func(t *testing.T) { testRegisterSigner(t, userSigners) })
 	t.Run("test signer exist in user", func(t *testing.T) { testSignerExistInUser(t, userSigners) })
@@ -107,9 +119,23 @@ func testGenerateToken(t *testing.T) {
 		Extra: "",
 	}
 
-	token1, err := jwtOAuthInstance.GenerateToken(context.Background(), pl1)
+	_, err := jwtOAuthInstance.GenerateToken(adminCtx, pl1)
+	assert.NotNil(t, err)
+	require.Contains(t, err.Error(), "token must be based on an existing user")
+
+	createUserReq := &CreateUserRequest{
+		Name:  "test-token-01",
+		State: 0,
+	}
+	resp, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
 	assert.Nil(t, err)
-	assert.Equal(t, 3, len(strings.Split(token1, ".")))
+	assert.Equal(t, "test-token-01", resp.Name)
+	token, err := jwtOAuthInstance.GenerateToken(adminCtx, pl1)
+	assert.Nil(t, err)
+	assert.Equal(t, 3, len(strings.Split(token, ".")))
+
+	_, err = jwtOAuthInstance.GenerateToken(signCtx, pl1)
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testVerifyToken(t *testing.T) {
@@ -123,19 +149,30 @@ func testVerifyToken(t *testing.T) {
 		Perm:  "admin",
 		Extra: "",
 	}
-	ctx := context.Background()
-	token1, err := jwtOAuthInstance.GenerateToken(ctx, pl1)
+
+	createUserReq := &CreateUserRequest{
+		Name:  "test-token-01",
+		State: 0,
+	}
+	resp, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
+	assert.Nil(t, err)
+	assert.Equal(t, "test-token-01", resp.Name)
+	token, err := jwtOAuthInstance.GenerateToken(adminCtx, pl1)
 	assert.Nil(t, err)
 
 	// Verify a valid token
-	payload1, err := jwtOAuthInstance.Verify(ctx, token1)
+	payload1, err := jwtOAuthInstance.Verify(readCtx, token)
 	assert.Nil(t, err)
 	assert.True(t, reflect.DeepEqual(payload1, pl1))
 
 	// Try to verify an invalid token
 	invalidToken := "I'm just an invalid token"
-	_, err = jwtOAuthInstance.Verify(ctx, invalidToken)
+	_, err = jwtOAuthInstance.Verify(readCtx, invalidToken)
 	assert.NotNil(t, err)
+
+	// with ctx no perm
+	_, err = jwtOAuthInstance.Verify(context.Background(), token)
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
 }
 
 func testGetToken(t *testing.T) {
@@ -149,19 +186,33 @@ func testGetToken(t *testing.T) {
 		Perm:  "admin",
 		Extra: "",
 	}
-	ctx := context.Background()
-	token1, err := jwtOAuthInstance.GenerateToken(ctx, pl1)
+
+	createUserReq := &CreateUserRequest{
+		Name:  "test-token-01",
+		State: 0,
+	}
+	resp, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
+	assert.Nil(t, err)
+	assert.Equal(t, "test-token-01", resp.Name)
+	token, err := jwtOAuthInstance.GenerateToken(adminCtx, pl1)
 	assert.Nil(t, err)
 
 	// Get token
-	tokenInfo1, err := jwtOAuthInstance.GetToken(ctx, token1)
+	tokenInfo1, err := jwtOAuthInstance.GetToken(adminCtx, token)
 	assert.Nil(t, err)
 	assert.Equal(t, pl1.Name, tokenInfo1.Name)
 	assert.Equal(t, pl1.Perm, tokenInfo1.Perm)
 	// Try to get invalid token
 	invalidToken := "I'm just an invalid token"
-	_, err = jwtOAuthInstance.GetToken(ctx, invalidToken)
+	_, err = jwtOAuthInstance.GetToken(adminCtx, invalidToken)
 	assert.NotNil(t, err)
+
+	// with ctx no perm
+	_, err = jwtOAuthInstance.GetToken(context.Background(), token)
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+
+	_, err = jwtOAuthInstance.GetToken(signCtx, invalidToken)
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testGetTokenByName(t *testing.T) {
@@ -175,17 +226,38 @@ func testGetTokenByName(t *testing.T) {
 		Perm:  "admin",
 		Extra: "",
 	}
-	ctx := context.Background()
-	token1, err := jwtOAuthInstance.GenerateToken(ctx, pl1)
-	assert.Nil(t, err)
 
-	// Get token by name
-	tokenInfoList1, err := jwtOAuthInstance.GetTokenByName(ctx, "test-token-01")
+	createUserReq := &CreateUserRequest{
+		Name:  "test-token-01",
+		State: 0,
+	}
+	resp, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
 	assert.Nil(t, err)
-	assert.Equal(t, 1, len(tokenInfoList1))
-	assert.Equal(t, token1, tokenInfoList1[0].Token)
+	assert.Equal(t, "test-token-01", resp.Name)
+	token, err := jwtOAuthInstance.GenerateToken(adminCtx, pl1)
+	assert.Nil(t, err)
+	userCtx := ctxWithUserNameAndAdminPerm(pl1.Name)
+
+	validPermTest := func(ctx context.Context) {
+		// Get token by name
+		tokenInfoList1, err := jwtOAuthInstance.GetTokenByName(ctx, "test-token-01")
+		assert.Nil(t, err)
+		assert.Equal(t, 1, len(tokenInfoList1))
+		assert.Equal(t, token, tokenInfoList1[0].Token)
+
+	}
+	invalidPermTest := func(ctx context.Context, expect error) {
+		_, err := jwtOAuthInstance.GetTokenByName(ctx, "test-token-01")
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+
+	validPermTest(adminCtx)
+	validPermTest(userCtx)
+	invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
+
 	// Try to get token by wrong name
-	tokenInfoInvalid, err := jwtOAuthInstance.GetTokenByName(ctx, "invalid_name")
+	tokenInfoInvalid, err := jwtOAuthInstance.GetTokenByName(adminCtx, "invalid_name")
 	assert.Nil(t, err)
 	assert.Equal(t, 0, len(tokenInfoInvalid))
 }
@@ -206,27 +278,48 @@ func testTokenList(t *testing.T) {
 		Perm:  "admin",
 		Extra: "",
 	}
-	ctx := context.Background()
-	_, err := jwtOAuthInstance.GenerateToken(ctx, pl1)
+
+	createUserReq := &CreateUserRequest{
+		Name:  "test-token-01",
+		State: 0,
+	}
+	resp, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
 	assert.Nil(t, err)
-	_, err = jwtOAuthInstance.GenerateToken(ctx, pl2)
+	assert.Equal(t, "test-token-01", resp.Name)
+	_, err = jwtOAuthInstance.GenerateToken(adminCtx, pl1)
+	assert.Nil(t, err)
+	createUserReq = &CreateUserRequest{
+		Name:  "test-token-02",
+		State: 0,
+	}
+	resp, err = jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
+	assert.Nil(t, err)
+	assert.Equal(t, "test-token-02", resp.Name)
+	_, err = jwtOAuthInstance.GenerateToken(adminCtx, pl2)
 	assert.Nil(t, err)
 
-	allTokenInfos, err := jwtOAuthInstance.Tokens(ctx, 0, 2)
+	allTokenInfos, err := jwtOAuthInstance.Tokens(adminCtx, 0, 2)
 	assert.Nil(t, err)
 	assert.Equal(t, 2, len(allTokenInfos))
 	// with skip or limit
-	allTokenInfos, err = jwtOAuthInstance.Tokens(ctx, 1, 10)
+	allTokenInfos, err = jwtOAuthInstance.Tokens(adminCtx, 1, 10)
 	assert.Nil(t, err)
 	assert.Equal(t, 1, len(allTokenInfos))
 
-	allTokenInfos, err = jwtOAuthInstance.Tokens(ctx, 0, 1)
+	allTokenInfos, err = jwtOAuthInstance.Tokens(adminCtx, 0, 1)
 	assert.Nil(t, err)
 	assert.Equal(t, 1, len(allTokenInfos))
 
-	allTokenInfos, err = jwtOAuthInstance.Tokens(ctx, 2, 10)
+	allTokenInfos, err = jwtOAuthInstance.Tokens(adminCtx, 2, 10)
 	assert.Nil(t, err)
 	assert.Equal(t, 0, len(allTokenInfos))
+
+	// with ctx no perm
+	_, err = jwtOAuthInstance.Tokens(context.Background(), 0, 2)
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+
+	_, err = jwtOAuthInstance.Tokens(signCtx, 0, 2)
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testRemoveAndRecoverToken(t *testing.T) {
@@ -237,41 +330,69 @@ func testRemoveAndRecoverToken(t *testing.T) {
 	// First generate a token
 	pl1 := &JWTPayload{
 		Name:  "test-token-01",
-		Perm:  "admin",
+		Perm:  "read",
 		Extra: "",
 	}
-	ctx := context.Background()
-	token1, err := jwtOAuthInstance.GenerateToken(ctx, pl1)
+
+	createUserReq := &CreateUserRequest{
+		Name:  "test-token-01",
+		State: 0,
+	}
+	resp, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
+	assert.Nil(t, err)
+	assert.Equal(t, "test-token-01", resp.Name)
+	token, err := jwtOAuthInstance.GenerateToken(adminCtx, pl1)
 	assert.Nil(t, err)
 
-	// token is usable.
-	err = jwtOAuthInstance.RecoverToken(ctx, token1)
-	assert.Error(t, err)
+	validPermTest := func(ctx context.Context) {
 
-	// Remove a token
-	err = jwtOAuthInstance.RemoveToken(ctx, token1)
-	assert.Nil(t, err)
+		// token is usable.
+		err = jwtOAuthInstance.RecoverToken(ctx, token)
+		assert.Error(t, err)
 
-	_, err = jwtOAuthInstance.Verify(ctx, token1)
-	assert.NotNil(t, err)
+		// Remove a token
+		err = jwtOAuthInstance.RemoveToken(ctx, token)
+		assert.Nil(t, err)
 
-	tokenInfoList1, err := jwtOAuthInstance.GetTokenByName(ctx, "test-token-01")
-	assert.Nil(t, err)
-	assert.Equal(t, 0, len(tokenInfoList1))
+		_, err = jwtOAuthInstance.Verify(ctx, token)
+		assert.NotNil(t, err)
 
-	// Recover a token
-	err = jwtOAuthInstance.RecoverToken(ctx, token1)
-	assert.Nil(t, err)
-	payload1, err := jwtOAuthInstance.Verify(ctx, token1)
-	assert.Nil(t, err)
-	assert.True(t, reflect.DeepEqual(payload1, pl1))
-	allTokenInfos, err := jwtOAuthInstance.Tokens(ctx, 0, 2)
-	assert.Nil(t, err)
-	assert.Equal(t, 1, len(allTokenInfos))
+		tokenInfoList1, err := jwtOAuthInstance.GetTokenByName(ctx, "test-token-01")
+		assert.Nil(t, err)
+		assert.Equal(t, 0, len(tokenInfoList1))
+
+		// Recover a token
+		err = jwtOAuthInstance.RecoverToken(ctx, token)
+		assert.Nil(t, err)
+		payload1, err := jwtOAuthInstance.Verify(ctx, token)
+		assert.Nil(t, err)
+		assert.True(t, reflect.DeepEqual(payload1, pl1))
+		allTokenInfos, err := jwtOAuthInstance.Tokens(adminCtx, 0, 2)
+		assert.Nil(t, err)
+		assert.Equal(t, 1, len(allTokenInfos))
+	}
+
+	invalidPermTest := func(ctx context.Context, expect error) {
+		// Remove a token
+		err := jwtOAuthInstance.RemoveToken(ctx, token)
+		assert.Equal(t, expect, errors.Unwrap(err))
+
+		// Recover a token
+		err = jwtOAuthInstance.RecoverToken(ctx, token)
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+
+	userCtx := ctxWithUserNameAndAdminPerm(pl1.Name)
+
+	validPermTest(userCtx)
+	validPermTest(adminCtx)
+
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
+	invalidPermTest(signCtx, ErrorUsernameNotFound)
 }
 
 func createUsers(t *testing.T, userMiners map[string][]string) {
-	ctx := context.Background()
+	ctx := adminCtx
 	// Create 3 users
 	for userName := range userMiners {
 		createUserReq := &CreateUserRequest{
@@ -290,8 +411,7 @@ func testCreateUser(t *testing.T, userMiners map[string][]string) {
 	setup(&cfg, t)
 	defer shutdown(&cfg, t)
 
-	ctx := context.Background()
-
+	// with adminCtx admin perm
 	existUserName := "test_user_001"
 	comment := "test comment"
 	// Create 3 users
@@ -301,14 +421,21 @@ func testCreateUser(t *testing.T, userMiners map[string][]string) {
 			Comment: &comment,
 			State:   0,
 		}
-		resp, err := jwtOAuthInstance.CreateUser(ctx, createUserReq)
+		resp, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
 		assert.Nil(t, err)
 		assert.Equal(t, userName, resp.Name)
 		assert.Equal(t, "test comment", resp.Comment)
 	}
 	// Create duplicate user
-	_, err := jwtOAuthInstance.CreateUser(ctx, &CreateUserRequest{Name: existUserName})
+	_, err := jwtOAuthInstance.CreateUser(adminCtx, &CreateUserRequest{Name: existUserName})
 	assert.NotNil(t, err)
+
+	// with ctx no perm
+	_, err = jwtOAuthInstance.CreateUser(context.Background(), &CreateUserRequest{Name: "test_user_002"})
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+
+	_, err = jwtOAuthInstance.CreateUser(signCtx, &CreateUserRequest{Name: "test_user_002"})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testGetUser(t *testing.T, userMiners map[string][]string) {
@@ -320,18 +447,26 @@ func testGetUser(t *testing.T, userMiners map[string][]string) {
 	defer shutdown(&cfg, t)
 	createUsers(t, userMiners)
 
-	ctx := context.Background()
-	// HasUser
-	exist, err := jwtOAuthInstance.HasUser(ctx, &HasUserRequest{Name: existUserName})
-	assert.Nil(t, err)
-	assert.True(t, exist)
-	exist, err = jwtOAuthInstance.HasUser(ctx, &HasUserRequest{Name: invalidUserName})
+	validPermTest := func(ctx context.Context) {
+		u, err := jwtOAuthInstance.GetUser(ctx, &GetUserRequest{Name: existUserName})
+		assert.Nil(t, err)
+		assert.Equal(t, u.Name, existUserName)
+	}
+	invalidPermTest := func(ctx context.Context, expect error) {
+		_, err := jwtOAuthInstance.GetUser(ctx, &GetUserRequest{Name: existUserName})
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+
+	userCtx := ctxWithUserNameAndAdminPerm(existUserName)
+	validPermTest(adminCtx)
+	validPermTest(userCtx)
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
+	invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
+
+	exist, err := jwtOAuthInstance.HasUser(adminCtx, &HasUserRequest{Name: invalidUserName})
 	assert.Nil(t, err)
 	assert.False(t, exist)
 
-	u, err := jwtOAuthInstance.GetUser(ctx, &GetUserRequest{Name: existUserName})
-	assert.Nil(t, err)
-	assert.Equal(t, u.Name, existUserName)
 }
 
 func testVerifyUsers(t *testing.T, userMiners map[string][]string) {
@@ -345,9 +480,16 @@ func testVerifyUsers(t *testing.T, userMiners map[string][]string) {
 		usernames = append(usernames, key)
 	}
 
-	ctx := context.Background()
-	err := jwtOAuthInstance.VerifyUsers(ctx, &VerifyUsersReq{Names: usernames})
+	// with ctx admin perm
+	err := jwtOAuthInstance.VerifyUsers(adminCtx, &VerifyUsersReq{Names: usernames})
 	assert.Nil(t, err)
+
+	// with ctx no perm
+	err = jwtOAuthInstance.VerifyUsers(context.Background(), &VerifyUsersReq{Names: usernames})
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+
+	err = jwtOAuthInstance.VerifyUsers(signCtx, &VerifyUsersReq{Names: usernames})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testListUser(t *testing.T, userMiners map[string][]string) {
@@ -356,12 +498,26 @@ func testListUser(t *testing.T, userMiners map[string][]string) {
 	defer shutdown(&cfg, t)
 	createUsers(t, userMiners)
 
-	allUserInfos, err := jwtOAuthInstance.ListUsers(context.Background(), &ListUsersRequest{
+	// with ctx admin perm
+	allUserInfos, err := jwtOAuthInstance.ListUsers(adminCtx, &ListUsersRequest{
 		Page:  &core.Page{},
 		State: int(core.UserStateUndefined),
 	})
 	assert.Nil(t, err)
 	assert.Equal(t, 3, len(allUserInfos))
+
+	// with ctx no perm
+	_, err = jwtOAuthInstance.ListUsers(context.Background(), &ListUsersRequest{
+		Page:  &core.Page{},
+		State: int(core.UserStateUndefined),
+	})
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+
+	_, err = jwtOAuthInstance.ListUsers(signCtx, &ListUsersRequest{
+		Page:  &core.Page{},
+		State: int(core.UserStateUndefined),
+	})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testUpdateUser(t *testing.T, userMiners map[string][]string) {
@@ -378,17 +534,25 @@ func testUpdateUser(t *testing.T, userMiners map[string][]string) {
 		Name:    existUserName,
 		Comment: &comment,
 	}
-	ctx := context.Background()
-	err := jwtOAuthInstance.UpdateUser(ctx, updateUserReq)
+
+	// with adminCtx admin perm
+	err := jwtOAuthInstance.UpdateUser(adminCtx, updateUserReq)
 	assert.Nil(t, err)
 	// Then get this user
-	outPutUser1, err := jwtOAuthInstance.GetUser(ctx, &GetUserRequest{Name: existUserName})
+	outPutUser1, err := jwtOAuthInstance.GetUser(adminCtx, &GetUserRequest{Name: existUserName})
 	assert.Nil(t, err)
 	assert.Equal(t, "New Comment", outPutUser1.Comment)
 
 	// invalid user name
-	err = jwtOAuthInstance.UpdateUser(ctx, &UpdateUserRequest{})
+	err = jwtOAuthInstance.UpdateUser(adminCtx, &UpdateUserRequest{})
 	assert.Error(t, err)
+
+	// with ctx no perm
+	err = jwtOAuthInstance.UpdateUser(context.Background(), updateUserReq)
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+
+	err = jwtOAuthInstance.UpdateUser(signCtx, updateUserReq)
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testDeleteAndRecoverUser(t *testing.T, userMiners map[string][]string) {
@@ -400,16 +564,16 @@ func testDeleteAndRecoverUser(t *testing.T, userMiners map[string][]string) {
 	defer shutdown(&cfg, t)
 	createUsers(t, userMiners)
 
-	ginCtx := &gin.Context{}
+	// with ctx admin perm
 
 	// Delete User
-	err := jwtOAuthInstance.DeleteUser(ginCtx, &DeleteUserRequest{Name: existUserName})
+	err := jwtOAuthInstance.DeleteUser(adminCtx, &DeleteUserRequest{Name: existUserName})
 	assert.Nil(t, err)
 	// Then try to get this user
-	_, err = jwtOAuthInstance.GetUser(ginCtx, &GetUserRequest{Name: existUserName})
+	_, err = jwtOAuthInstance.GetUser(adminCtx, &GetUserRequest{Name: existUserName})
 	assert.NotNil(t, err)
 	// And also list users now
-	allUserInfos, err := jwtOAuthInstance.ListUsers(ginCtx, &ListUsersRequest{
+	allUserInfos, err := jwtOAuthInstance.ListUsers(adminCtx, &ListUsersRequest{
 		Page:  &core.Page{},
 		State: int(core.UserStateUndefined),
 	})
@@ -417,28 +581,39 @@ func testDeleteAndRecoverUser(t *testing.T, userMiners map[string][]string) {
 	assert.Equal(t, 2, len(allUserInfos))
 
 	// Try to delete non-existing users
-	err = jwtOAuthInstance.DeleteUser(ginCtx, &DeleteUserRequest{Name: invalidUserName})
+	err = jwtOAuthInstance.DeleteUser(adminCtx, &DeleteUserRequest{Name: invalidUserName})
 	assert.NotNil(t, err)
 
 	// Recover user
-	err = jwtOAuthInstance.RecoverUser(ginCtx, &RecoverUserRequest{Name: existUserName})
+	err = jwtOAuthInstance.RecoverUser(adminCtx, &RecoverUserRequest{Name: existUserName})
 	assert.Nil(t, err)
 	// Then get this user
-	outPutUser1, err := jwtOAuthInstance.GetUser(ginCtx, &GetUserRequest{Name: existUserName})
+	outPutUser1, err := jwtOAuthInstance.GetUser(adminCtx, &GetUserRequest{Name: existUserName})
 	assert.Nil(t, err)
 	assert.Equal(t, existUserName, outPutUser1.Name)
 
 	// Try to recover an invalid user
-	err = jwtOAuthInstance.RecoverUser(ginCtx, &RecoverUserRequest{Name: invalidUserName})
+	err = jwtOAuthInstance.RecoverUser(adminCtx, &RecoverUserRequest{Name: invalidUserName})
 	assert.NotNil(t, err)
 
 	// Try to recover a valid, but not deleted user
-	err = jwtOAuthInstance.RecoverUser(ginCtx, &RecoverUserRequest{Name: existUserName})
+	err = jwtOAuthInstance.RecoverUser(adminCtx, &RecoverUserRequest{Name: existUserName})
 	assert.NotNil(t, err)
+
+	// with ctx no perm
+	err = jwtOAuthInstance.DeleteUser(context.Background(), &DeleteUserRequest{Name: existUserName})
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+	err = jwtOAuthInstance.DeleteUser(signCtx, &DeleteUserRequest{Name: existUserName})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
+
+	err = jwtOAuthInstance.RecoverUser(context.Background(), &RecoverUserRequest{Name: existUserName})
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+	err = jwtOAuthInstance.RecoverUser(signCtx, &RecoverUserRequest{Name: existUserName})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func addUsersAndMiners(t *testing.T, userMiners map[string][]string) {
-	ctx := context.Background()
+	ctx := adminCtx
 	for userName, miners := range userMiners {
 		createUserReq := &CreateUserRequest{
 			Name:  userName,
@@ -449,9 +624,11 @@ func addUsersAndMiners(t *testing.T, userMiners map[string][]string) {
 		// Add miners
 		openMining := true
 		for _, minerID := range miners {
+			mAddr, err := address.NewFromString(minerID)
+			assert.Nil(t, err)
 			ifCreate, err := jwtOAuthInstance.UpsertMiner(ctx, &UpsertMinerReq{
 				User:       userName,
-				Miner:      minerID,
+				Miner:      mAddr,
 				OpenMining: &openMining,
 			})
 			assert.Nil(t, err)
@@ -466,22 +643,29 @@ func testUpsertMiner(t *testing.T, userMiners map[string][]string) {
 	defer shutdown(&cfg, t)
 	addUsersAndMiners(t, userMiners)
 
-	ctx := context.Background()
+	// with adminCtx
 	// error signer address
-	_, _ = jwtOAuthInstance.CreateUser(ctx, &CreateUserRequest{
+	_, _ = jwtOAuthInstance.CreateUser(adminCtx, &CreateUserRequest{
 		Name:  "user_01",
 		State: 1,
 	})
-	isCreate, err := jwtOAuthInstance.UpsertMiner(ctx, &UpsertMinerReq{User: "user_01", Miner: "f01034"})
+	mAddr, _ := address.NewFromString("f01034")
+	isCreate, err := jwtOAuthInstance.UpsertMiner(adminCtx, &UpsertMinerReq{User: "user_01", Miner: mAddr})
 	assert.Nil(t, err)
 	assert.True(t, isCreate)
 
-	_, err = jwtOAuthInstance.UpsertMiner(ctx, &UpsertMinerReq{
+	mAddr, err = address.NewFromString("f1mpvdqt2acgihevibd4greavlsfn3dfph5sckc2a")
+	assert.Nil(t, err)
+	_, err = jwtOAuthInstance.UpsertMiner(adminCtx, &UpsertMinerReq{
 		User:  "user_01",
-		Miner: "f1mpvdqt2acgihevibd4greavlsfn3dfph5sckc2a",
+		Miner: mAddr,
 	})
 	assert.NotNil(t, err)
 	require.Contains(t, err.Error(), "invalid protocol type")
+
+	// with signCtx
+	_, err = jwtOAuthInstance.UpsertMiner(signCtx, &UpsertMinerReq{User: "user_01", Miner: mAddr})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testListMiner(t *testing.T, userMiners map[string][]string) {
@@ -492,43 +676,76 @@ func testListMiner(t *testing.T, userMiners map[string][]string) {
 
 	validUser1 := "test_user_001"
 	user1Miners := []string{"t01000", "t01002", "t01003"}
-	ctx := context.Background()
-	// List miners
-	resp, err := jwtOAuthInstance.ListMiners(ctx, &ListMinerReq{User: validUser1})
-	assert.Nil(t, err)
-	assert.Equal(t, len(user1Miners), len(resp))
-	sort.Slice(resp, func(i, j int) bool { return resp[i].Miner < resp[j].Miner })
-	for i := 0; i < len(user1Miners); i++ {
-		assert.Equal(t, user1Miners[i], resp[i].Miner)
-		assert.Equal(t, validUser1, resp[i].User)
-		assert.Equal(t, true, resp[i].OpenMining)
+
+	validPermTest := func(ctx context.Context) {
+		// List miners
+		resp, err := jwtOAuthInstance.ListMiners(ctx, &ListMinerReq{User: validUser1})
+		assert.Nil(t, err)
+		assert.Equal(t, len(user1Miners), len(resp))
+		sort.Slice(resp, func(i, j int) bool { return resp[i].Miner.String() < resp[j].Miner.String() })
+		for i := 0; i < len(user1Miners); i++ {
+			assert.Equal(t, user1Miners[i], resp[i].Miner.String())
+			assert.Equal(t, validUser1, resp[i].User)
+			assert.Equal(t, true, resp[i].OpenMining)
+		}
 	}
+	invalidPermTest := func(ctx context.Context, expect error) {
+		_, err := jwtOAuthInstance.ListMiners(ctx, &ListMinerReq{User: validUser1})
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+
+	userCtx := ctxWithUserNameAndAdminPerm(validUser1)
+	validPermTest(userCtx)
+	validPermTest(adminCtx)
+
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
+	invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
 }
 
-func testHasMiner(t *testing.T, userMiners map[string][]string) {
+func testMinerExistInMiner(t *testing.T, userMiners map[string][]string) {
 	cfg := config.DBConfig{Type: "badger"}
 	setup(&cfg, t)
 	defer shutdown(&cfg, t)
 	addUsersAndMiners(t, userMiners)
 
-	ctx := context.Background()
+	validPermTest := func(ctx context.Context) {
+		mAddr1, err := address.NewFromString("t01000")
+		assert.Nil(t, err)
+		// Miner Exist In Account
+		exist, err := jwtOAuthInstance.MinerExistInUser(ctx, &MinerExistInUserRequest{Miner: mAddr1, User: "test_user_001"})
+		assert.Nil(t, err)
+		assert.True(t, exist)
+
+		mAddr4, err := address.NewFromString("t01004")
+		assert.Nil(t, err)
+		exist, err = jwtOAuthInstance.MinerExistInUser(ctx, &MinerExistInUserRequest{Miner: mAddr4, User: "test_user_001"})
+		assert.Nil(t, err)
+		assert.False(t, exist)
+	}
+	invalidPermTest := func(ctx context.Context, expect error) {
+		mAddr, _ := address.NewFromString("t01000")
+		// Miner Exist In Account
+		_, err := jwtOAuthInstance.MinerExistInUser(ctx, &MinerExistInUserRequest{Miner: mAddr, User: "test_user_001"})
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+
+	userCtx := ctxWithUserNameAndAdminPerm("test_user_001")
+	validPermTest(userCtx)
+	validPermTest(adminCtx)
+
+	invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
 
 	// Has Miner
-	has, err := jwtOAuthInstance.HasMiner(ctx, &HasMinerRequest{Miner: "t01000"})
+	mAddr, _ := address.NewFromString("t01000")
+	has, err := jwtOAuthInstance.HasMiner(adminCtx, &HasMinerRequest{Miner: mAddr})
 	assert.Nil(t, err)
 	assert.True(t, has)
-
-	// Miner Exist In Account
-	exist, err := jwtOAuthInstance.MinerExistInUser(ctx, &MinerExistInUserRequest{Miner: "t01000", User: "test_user_001"})
-	assert.Nil(t, err)
-	assert.True(t, exist)
-
-	exist, err = jwtOAuthInstance.MinerExistInUser(ctx, &MinerExistInUserRequest{Miner: "t01000", User: "test_user_002"})
-	assert.Nil(t, err)
-	assert.False(t, exist)
-
-	_, err = jwtOAuthInstance.HasMiner(ctx, &HasMinerRequest{Miner: "invalid address"})
-	assert.Error(t, err)
+	invalidMiner, _ := address.NewFromString("t01014")
+	has, _ = jwtOAuthInstance.HasMiner(adminCtx, &HasMinerRequest{Miner: invalidMiner})
+	assert.False(t, has)
+	_, err = jwtOAuthInstance.HasMiner(signCtx, &HasMinerRequest{Miner: mAddr})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testGetUserByMiner(t *testing.T, userMiners map[string][]string) {
@@ -537,27 +754,32 @@ func testGetUserByMiner(t *testing.T, userMiners map[string][]string) {
 	defer shutdown(&cfg, t)
 	addUsersAndMiners(t, userMiners)
 
-	ctx := context.Background()
+	// with ctx admin perm
 	// Get User By Miner
 	validUser1 := "test_user_001"
-	user1Miners := []string{"t01000", "t01002", "t01003"}
-	userInfo, err := jwtOAuthInstance.GetUserByMiner(ctx, &GetUserByMinerRequest{
-		Miner: user1Miners[1],
+	miner := "t01002"
+	mAddr, err := address.NewFromString(miner)
+	assert.Nil(t, err)
+	userInfo, err := jwtOAuthInstance.GetUserByMiner(adminCtx, &GetUserByMinerRequest{
+		Miner: mAddr,
 	})
 	assert.Nil(t, err)
 	assert.Equal(t, validUser1, userInfo.Name)
 
-	// invalid miner address
-	_, err = jwtOAuthInstance.GetUserByMiner(ctx, &GetUserByMinerRequest{
-		Miner: "invalid address",
+	// miner address not exist
+	mAddr, err = address.NewFromString("f01989787")
+	assert.Nil(t, err)
+	_, err = jwtOAuthInstance.GetUserByMiner(adminCtx, &GetUserByMinerRequest{
+		Miner: mAddr,
 	})
 	assert.Error(t, err)
 
-	// miner address not exist
-	_, err = jwtOAuthInstance.GetUserByMiner(ctx, &GetUserByMinerRequest{
-		Miner: "f01989787",
+	// with ctx no perm
+	// Get User By Mine
+	_, err = jwtOAuthInstance.GetUserByMiner(signCtx, &GetUserByMinerRequest{
+		Miner: mAddr,
 	})
-	assert.Error(t, err)
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testDeleteMiner(t *testing.T, userMiners map[string][]string) {
@@ -566,25 +788,46 @@ func testDeleteMiner(t *testing.T, userMiners map[string][]string) {
 	defer shutdown(&cfg, t)
 	addUsersAndMiners(t, userMiners)
 
+	user1 := "test_user_001"
 	user1Miners := []string{"t01000", "t01002", "t01003"}
 	invalidMiner := "t02000"
-	ctx := context.Background()
-	// Delete miner
-	deleted, err := jwtOAuthInstance.DelMiner(ctx, &DelMinerReq{Miner: user1Miners[0]})
-	assert.Nil(t, err)
-	assert.True(t, deleted)
-	// Then get this miner
-	has, err := jwtOAuthInstance.HasMiner(ctx, &HasMinerRequest{Miner: user1Miners[0]})
-	assert.Nil(t, err)
-	assert.False(t, has)
-	// Try to get user by this miner
-	_, err = jwtOAuthInstance.GetUserByMiner(ctx, &GetUserByMinerRequest{
-		Miner: user1Miners[0],
-	})
-	assert.NotNil(t, err)
+
+	validPermTest := func(ctx context.Context) {
+		// Delete miner
+		mAddr, err := address.NewFromString(user1Miners[0])
+		assert.Nil(t, err)
+		deleted, err := jwtOAuthInstance.DelMiner(ctx, &DelMinerReq{Miner: mAddr})
+		assert.Nil(t, err)
+		assert.True(t, deleted)
+		// Then get this miner
+		has, err := jwtOAuthInstance.HasMiner(adminCtx, &HasMinerRequest{Miner: mAddr})
+		assert.Nil(t, err)
+		assert.False(t, has)
+
+		ok, err := jwtOAuthInstance.UpsertMiner(adminCtx, &UpsertMinerReq{
+			User:  user1,
+			Miner: mAddr,
+		})
+		assert.Nil(t, err)
+		assert.False(t, ok)
+
+	}
+	invalidPermTest := func(ctx context.Context, expect error) {
+		mAddr, _ := address.NewFromString(user1Miners[0])
+		// Delete miner
+		_, err := jwtOAuthInstance.DelMiner(ctx, &DelMinerReq{Miner: mAddr})
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+	userCtx := ctxWithUserNameAndAdminPerm(user1)
+	validPermTest(userCtx)
+	validPermTest(adminCtx)
+
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
+	invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
 
 	// Delete an invalid miner
-	deleted, err = jwtOAuthInstance.DelMiner(ctx, &DelMinerReq{Miner: invalidMiner})
+	mAddr, _ := address.NewFromString(invalidMiner)
+	deleted, err := jwtOAuthInstance.DelMiner(adminCtx, &DelMinerReq{Miner: mAddr})
 	assert.Nil(t, err)
 	assert.False(t, deleted)
 }
@@ -596,15 +839,30 @@ func addUsersAndSigners(t *testing.T, userSigners map[string][]string) {
 			State: 1,
 		}
 
-		ctx := context.Background()
+		// with adminCtx admin perm
 		// Create users.
-		_, _ = jwtOAuthInstance.CreateUser(ctx, createUserReq)
+		_, err := jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
+		assert.NoError(t, err)
 		// Add Signer
-		err := jwtOAuthInstance.RegisterSigners(ctx, &RegisterSignersReq{
+		signerAddrs := make([]address.Address, 0)
+		for _, signer := range signers {
+			signerAddr, err := address.NewFromString(signer)
+			assert.Nil(t, err)
+			signerAddrs = append(signerAddrs, signerAddr)
+		}
+		err = jwtOAuthInstance.RegisterSigners(adminCtx, &RegisterSignersReq{
 			User:    userName,
-			Signers: signers,
+			Signers: signerAddrs,
 		})
 		assert.Nil(t, err)
+
+		// with ctx no perm
+		// Add Signer
+		err = jwtOAuthInstance.RegisterSigners(signCtx, &RegisterSignersReq{
+			User:    userName,
+			Signers: signerAddrs,
+		})
+		assert.True(t, errors.Is(err, ErrorPermissionDeny))
 	}
 }
 
@@ -616,24 +874,19 @@ func testRegisterSigner(t *testing.T, userSigners map[string][]string) {
 	addUsersAndSigners(t, userSigners)
 
 	// error signer address
-	ctx := context.Background()
+	ctx := adminCtx
 	_, _ = jwtOAuthInstance.CreateUser(ctx, &CreateUserRequest{
 		Name:  "user_01",
 		State: 1,
 	})
-	err := jwtOAuthInstance.RegisterSigners(ctx, &RegisterSignersReq{
+	sAddr, err := address.NewFromString("f0128788")
+	assert.Nil(t, err)
+	err = jwtOAuthInstance.RegisterSigners(ctx, &RegisterSignersReq{
 		User:    "user_01",
-		Signers: []string{"f0128788"},
+		Signers: []address.Address{sAddr},
 	})
 	assert.NotNil(t, err)
 	require.Contains(t, err.Error(), "invalid protocol type")
-
-	err = jwtOAuthInstance.RegisterSigners(ctx, &RegisterSignersReq{
-		User:    "user_01",
-		Signers: []string{"128788"},
-	})
-	assert.NotNil(t, err)
-	require.Contains(t, err.Error(), "invalid signer address")
 }
 
 func testSignerExistInUser(t *testing.T, userSigners map[string][]string) {
@@ -642,17 +895,37 @@ func testSignerExistInUser(t *testing.T, userSigners map[string][]string) {
 	defer shutdown(&cfg, t)
 
 	addUsersAndSigners(t, userSigners)
-
-	ctx := context.Background()
+	// test every user
 	for user, signers := range userSigners {
-		for _, signer := range signers {
-			bExist, err := jwtOAuthInstance.SignerExistInUser(ctx, &SignerExistInUserReq{
-				User:   user,
-				Signer: signer,
-			})
-			assert.Nil(t, err)
-			assert.True(t, bExist)
+
+		validPermTest := func(ctx context.Context) {
+			for _, signer := range signers {
+				sAddr, err := address.NewFromString(signer)
+				assert.Nil(t, err)
+				bExist, err := jwtOAuthInstance.SignerExistInUser(ctx, &SignerExistInUserReq{
+					User:   user,
+					Signer: sAddr,
+				})
+				assert.Nil(t, err)
+				assert.True(t, bExist)
+			}
 		}
+		invalidPermTest := func(ctx context.Context, expect error) {
+			for _, signer := range signers {
+				sAddr, _ := address.NewFromString(signer)
+				_, err := jwtOAuthInstance.SignerExistInUser(ctx, &SignerExistInUserReq{
+					User:   user,
+					Signer: sAddr,
+				})
+				assert.Equal(t, expect, errors.Unwrap(err))
+			}
+		}
+
+		userCtx := ctxWithUserNameAndAdminPerm(user)
+		validPermTest(userCtx)
+		validPermTest(adminCtx)
+		invalidPermTest(context.Background(), ErrorUsernameNotFound)
+		invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
 	}
 }
 
@@ -664,13 +937,27 @@ func testListSigner(t *testing.T, userSigners map[string][]string) {
 
 	validUser1 := "test_user_001"
 	user1Signers := userSigners[validUser1]
-	// List miners
-	resp, err := jwtOAuthInstance.ListSigner(context.Background(), &ListSignerReq{User: validUser1})
-	assert.Nil(t, err)
-	assert.Equal(t, len(user1Signers), len(resp))
-	for _, signer := range resp {
-		require.Contains(t, user1Signers, signer.Signer)
+
+	validPermTest := func(ctx context.Context) {
+		// List miners
+		resp, err := jwtOAuthInstance.ListSigner(adminCtx, &ListSignerReq{User: validUser1})
+		assert.Nil(t, err)
+		assert.Equal(t, len(user1Signers), len(resp))
+		for _, signer := range resp {
+			require.Contains(t, user1Signers, signer.Signer.String())
+		}
 	}
+	invalidPermTest := func(ctx context.Context, expect error) {
+		// List miners
+		_, err := jwtOAuthInstance.ListSigner(ctx, &ListSignerReq{User: validUser1})
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+
+	userCtx := ctxWithUserNameAndAdminPerm(validUser1)
+	validPermTest(userCtx)
+	validPermTest(adminCtx)
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
+	invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
 }
 
 func testHasSigner(t *testing.T, userSigners map[string][]string) {
@@ -678,14 +965,24 @@ func testHasSigner(t *testing.T, userSigners map[string][]string) {
 	setup(&cfg, t)
 	defer shutdown(&cfg, t)
 	addUsersAndSigners(t, userSigners)
-	ctx := context.Background()
 
-	has, err := jwtOAuthInstance.HasSigner(ctx, &HasSignerReq{Signer: "t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua"})
+	sAddr, err := address.NewFromString("t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua")
+	assert.Nil(t, err)
+	// with adminCtx
+	has, err := jwtOAuthInstance.HasSigner(adminCtx, &HasSignerReq{Signer: sAddr})
 	assert.Nil(t, err)
 	assert.True(t, has)
 
-	has, err = jwtOAuthInstance.HasSigner(ctx, &HasSignerReq{Signer: "f3r72mrymha6wrtb6dzynkzjbnl572az27ddbiq3aovj3d235h2jjgsya4afbf3d37vzfbtsy3dssfnitnhklq"})
+	sAddr, err = address.NewFromString("f3r72mrymha6wrtb6dzynkzjbnl572az27ddbiq3aovj3d235h2jjgsya4afbf3d37vzfbtsy3dssfnitnhklq")
 	assert.Nil(t, err)
+	has, err = jwtOAuthInstance.HasSigner(adminCtx, &HasSignerReq{Signer: sAddr})
+	assert.Nil(t, err)
+	assert.False(t, has)
+
+	// with signCtx
+	sAddr, _ = address.NewFromString("t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua")
+	has, err = jwtOAuthInstance.HasSigner(signCtx, &HasSignerReq{Signer: sAddr})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 	assert.False(t, has)
 }
 
@@ -695,10 +992,12 @@ func testGetUserBySigner(t *testing.T, userSigners map[string][]string) {
 	defer shutdown(&cfg, t)
 	addUsersAndSigners(t, userSigners)
 
+	// with adminCtx
 	// Get User By Signer
-	signer := "t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua"
-	users, err := jwtOAuthInstance.GetUserBySigner(context.Background(), &GetUserBySignerReq{
-		Signer: signer,
+	sAddr, err := address.NewFromString("t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua")
+	assert.Nil(t, err)
+	users, err := jwtOAuthInstance.GetUserBySigner(adminCtx, &GetUserBySignerReq{
+		Signer: sAddr,
 	})
 
 	names := make([]string, len(users))
@@ -709,6 +1008,12 @@ func testGetUserBySigner(t *testing.T, userSigners map[string][]string) {
 	assert.Nil(t, err)
 	require.Contains(t, names, "test_user_001")
 	require.Contains(t, names, "test_user_003")
+
+	// with signCtx
+	_, err = jwtOAuthInstance.GetUserBySigner(signCtx, &GetUserBySignerReq{
+		Signer: sAddr,
+	})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testUnregisterSigner(t *testing.T, userSigners map[string][]string) {
@@ -718,21 +1023,30 @@ func testUnregisterSigner(t *testing.T, userSigners map[string][]string) {
 	addUsersAndSigners(t, userSigners)
 
 	username := "test_user_001"
-	signer := "t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua"
-	ctx := context.Background()
-	err := jwtOAuthInstance.UnregisterSigners(ctx, &UnregisterSignersReq{
-		Signers: []string{signer},
-		User:    username,
-	})
-
+	sAddr, err := address.NewFromString("t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua")
 	assert.Nil(t, err)
 
-	bExist, err := jwtOAuthInstance.SignerExistInUser(ctx, &SignerExistInUserReq{
-		Signer: signer,
+	// with adminCtx admin perm
+	adminCtx := adminCtx
+	err = jwtOAuthInstance.UnregisterSigners(adminCtx, &UnregisterSignersReq{
+		Signers: []address.Address{sAddr},
+		User:    username,
+	})
+	assert.Nil(t, err)
+
+	bExist, err := jwtOAuthInstance.SignerExistInUser(adminCtx, &SignerExistInUserReq{
+		Signer: sAddr,
 		User:   username,
 	})
 	assert.Nil(t, err)
 	assert.False(t, bExist)
+
+	// with ctx sign perm
+	err = jwtOAuthInstance.UnregisterSigners(signCtx, &UnregisterSignersReq{
+		Signers: []address.Address{sAddr},
+		User:    username,
+	})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func testDeleteSigner(t *testing.T, userSigners map[string][]string) {
@@ -742,30 +1056,51 @@ func testDeleteSigner(t *testing.T, userSigners map[string][]string) {
 	addUsersAndSigners(t, userSigners)
 
 	// Delete signer
-	signer := "t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua"
-	ctx := context.Background()
-	deleted, err := jwtOAuthInstance.DelSigner(ctx, &DelSignerReq{Signer: signer})
+	signer, err := address.NewFromString("t15rynkupqyfx5ebvaishg7duutwb5ooq2qpaikua")
 	assert.Nil(t, err)
-	assert.True(t, deleted)
+	signers := []address.Address{signer}
+	userName := "test_user_001"
+	validPermTest := func(ctx context.Context) {
+		deleted, err := jwtOAuthInstance.DelSigner(ctx, &DelSignerReq{Signer: signer})
+		assert.Nil(t, err)
+		assert.True(t, deleted)
 
-	// Then get this signer
-	has, err := jwtOAuthInstance.HasSigner(ctx, &HasSignerReq{Signer: signer})
-	assert.Nil(t, err)
-	assert.False(t, has)
+		// Then get this signer
+		has, err := jwtOAuthInstance.HasSigner(adminCtx, &HasSignerReq{Signer: signer})
+		assert.Nil(t, err)
+		assert.False(t, has)
+
+		// Add Signer back
+		err = jwtOAuthInstance.RegisterSigners(adminCtx, &RegisterSignersReq{
+			User:    userName,
+			Signers: signers,
+		})
+		assert.Nil(t, err)
+	}
+	invalidPermTest := func(ctx context.Context, expect error) {
+		_, err := jwtOAuthInstance.DelSigner(ctx, &DelSignerReq{Signer: signer})
+		assert.Equal(t, expect, errors.Unwrap(err))
+	}
+
+	userCtx := ctxWithUserNameAndAdminPerm(userName)
+	validPermTest(userCtx)
+	validPermTest(adminCtx)
+	invalidPermTest(context.Background(), ErrorUsernameNotFound)
+	invalidPermTest(signPermAndRandomUsername(), ErrorPermissionDeny)
 }
 
 func addUsersAndRateLimits(t *testing.T, userMiners map[string][]string, originLimits []*storage.UserRateLimit) {
-	ctx := context.Background()
+	// with adminCtx perm admin
 	// Create 3 users and add rate limits
 	for userName := range userMiners {
 		createUserReq := &CreateUserRequest{
 			Name:  userName,
 			State: 0,
 		}
-		_, _ = jwtOAuthInstance.CreateUser(ctx, createUserReq)
+		_, _ = jwtOAuthInstance.CreateUser(adminCtx, createUserReq)
 	}
 	for _, limit := range originLimits {
-		id, err := jwtOAuthInstance.UpsertUserRateLimit(ctx, &UpsertUserRateLimitReq{
+		id, err := jwtOAuthInstance.UpsertUserRateLimit(adminCtx, &UpsertUserRateLimitReq{
 			Id:       limit.Id,
 			Name:     limit.Name,
 			Service:  limit.Service,
@@ -775,6 +1110,17 @@ func addUsersAndRateLimits(t *testing.T, userMiners map[string][]string, originL
 		assert.Nil(t, err)
 		assert.Equal(t, limit.Id, id)
 	}
+
+	// with ctx no perm
+	_, err := jwtOAuthInstance.UpsertUserRateLimit(signCtx, &UpsertUserRateLimitReq{
+		Id:       originLimits[0].Id,
+		Name:     originLimits[0].Name,
+		Service:  originLimits[0].Service,
+		API:      originLimits[0].API,
+		ReqLimit: originLimits[0].ReqLimit,
+	})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
+
 }
 
 func testUpsertUserRateLimit(t *testing.T, userMiners map[string][]string, originLimits []*storage.UserRateLimit) {
@@ -790,16 +1136,28 @@ func testGetUserRateLimits(t *testing.T, userMiners map[string][]string, originL
 	defer shutdown(&cfg, t)
 	addUsersAndRateLimits(t, userMiners, originLimits)
 
-	// Test GetUserRateLimits
+	// with ctx admin perm
 	userName := originLimits[0].Name
 	existId := originLimits[0].Id
-	resp, err := jwtOAuthInstance.GetUserRateLimits(context.Background(), &GetUserRateLimitsReq{
+	resp, err := jwtOAuthInstance.GetUserRateLimits(adminCtx, &GetUserRateLimitsReq{
 		Id:   existId,
 		Name: userName,
 	})
 	assert.Nil(t, err)
 	assert.Equal(t, 1, len(resp))
 	assert.Equal(t, existId, resp[0].Id)
+
+	// with ctx no perm
+	_, err = jwtOAuthInstance.GetUserRateLimits(context.Background(), &GetUserRateLimitsReq{
+		Id:   existId,
+		Name: userName,
+	})
+	assert.Equal(t, ErrorPermissionNotFound, errors.Unwrap(err))
+
+	_ = jwtOAuthInstance.DelUserRateLimit(signCtx, &DelUserRateLimitReq{
+		Name: userName,
+		Id:   existId,
+	})
 }
 
 func testDeleteUserRateLimits(t *testing.T, userMiners map[string][]string, originLimits []*storage.UserRateLimit) {
@@ -812,26 +1170,32 @@ func testDeleteUserRateLimits(t *testing.T, userMiners map[string][]string, orig
 	userName := originLimits[0].Name
 	existId := originLimits[0].Id
 
-	ctx := context.Background()
-
-	err := jwtOAuthInstance.DelUserRateLimit(ctx, &DelUserRateLimitReq{
+	// with ctx admin perm
+	err := jwtOAuthInstance.DelUserRateLimit(adminCtx, &DelUserRateLimitReq{
 		Name: userName,
 		Id:   existId,
 	})
 	assert.Nil(t, err)
 	// Try to get it again
-	resp, err := jwtOAuthInstance.GetUserRateLimits(ctx, &GetUserRateLimitsReq{
+	resp, err := jwtOAuthInstance.GetUserRateLimits(adminCtx, &GetUserRateLimitsReq{
 		Id:   existId,
 		Name: userName,
 	})
 	assert.Nil(t, err)
 	assert.Equal(t, 0, len(resp))
 	// Try to delete again
-	err = jwtOAuthInstance.DelUserRateLimit(ctx, &DelUserRateLimitReq{
+	err = jwtOAuthInstance.DelUserRateLimit(adminCtx, &DelUserRateLimitReq{
 		Name: userName,
 		Id:   existId,
 	})
 	assert.NotNil(t, err)
+
+	// with ctx no perm
+	err = jwtOAuthInstance.DelUserRateLimit(signCtx, &DelUserRateLimitReq{
+		Name: userName,
+		Id:   existId,
+	})
+	assert.True(t, errors.Is(err, ErrorPermissionDeny))
 }
 
 func TestTokenDecode(t *testing.T) {
@@ -850,29 +1214,18 @@ func TestTokenDecode(t *testing.T) {
 
 func setup(cfg *config.DBConfig, t *testing.T) {
 	var err error
+	var dataPath string
 	if cfg.Type == "badger" {
-		if cfg.DSN, err = ioutil.TempDir("", "auth-datastore"); err != nil {
-			t.Fatal(err)
-		}
-		fmt.Printf("tmp badger store : %s\n", cfg.DSN)
+		dataPath = t.TempDir()
 	}
-	theStore, err := storage.NewStore(cfg, cfg.DSN)
+	theStore, err := storage.NewStore(cfg, dataPath)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	secret, err := config.RandSecret()
-	if err != nil {
-		t.Fatal(err)
-	}
-	sec, err := hex.DecodeString(hex.EncodeToString(secret))
-	if err != nil {
-		t.Fatal(err)
-	}
 	jwtOAuthInstance = &jwtOAuth{
-		secret: jwt.NewHS256(sec),
-		store:  theStore,
-		mp:     newMapper(),
+		store: theStore,
+		mp:    newMapper(),
 	}
 }
 
@@ -882,4 +1235,8 @@ func shutdown(cfg *config.DBConfig, t *testing.T) {
 	if err := os.RemoveAll(cfg.DSN); err != nil {
 		t.Fatal(err)
 	}
+}
+
+func ctxWithUserNameAndAdminPerm(userName string) context.Context {
+	return core.CtxWithName(adminCtx, userName)
 }
