@@ -4,15 +4,15 @@ import (
 	"net/http"
 	"path"
 
-	"github.com/filecoin-project/venus-auth/auth"
-	"github.com/filecoin-project/venus-auth/config"
-	"github.com/filecoin-project/venus-auth/log"
 	"github.com/gin-gonic/gin"
+	"github.com/ipfs-force-community/sophon-auth/auth"
+	"github.com/ipfs-force-community/sophon-auth/config"
+	"github.com/ipfs-force-community/sophon-auth/log"
 	"github.com/mitchellh/go-homedir"
 )
 
 func main() {
-	absoluteTmp := "~/.venus-auth"
+	absoluteTmp := "~/.sophon-auth"
 	dir, err := homedir.Expand(absoluteTmp)
 	if err != nil {
 		log.Printf("could not expand repo location error:%s", err)
@@ -30,19 +30,19 @@ func main() {
 		return
 	}
 	log.InitLog(cnf.Log)
-	app, err := auth.NewOAuthApp(cnf.Secret, dataPath, cnf.DB)
+	app, err := auth.NewOAuthApp(dataPath, cnf.DB)
 	if err != nil {
-		log.Fatalf("Failed to init venus-auth: %s", err)
+		log.Fatalf("Failed to init sophon-auth: %s", err)
 	}
 	router := auth.InitRouter(app)
 	server := &http.Server{
-		Addr:         ":8989",
+		Addr:         "127.0.0.1:8989",
 		Handler:      router,
 		ReadTimeout:  cnf.ReadTimeout,
 		WriteTimeout: cnf.WriteTimeout,
 		IdleTimeout:  cnf.IdleTimeout,
 	}
-	log.Infof("server start and listen on %s", cnf.Port)
+	log.Infof("server start and listen on %s", cnf.Listen)
 	server.ListenAndServe() //nolint
 	select {}
 }
