@@ -1,33 +1,24 @@
 package integrate
 
 import (
-	"io/ioutil"
 	"net/http/httptest"
 	"os"
 	"path"
 	"testing"
 
-	"github.com/filecoin-project/venus-auth/auth"
-	"github.com/filecoin-project/venus-auth/config"
-	"github.com/filecoin-project/venus-auth/log"
 	"github.com/gin-gonic/gin"
+	"github.com/ipfs-force-community/sophon-auth/auth"
+	"github.com/ipfs-force-community/sophon-auth/config"
+	"github.com/ipfs-force-community/sophon-auth/log"
 	"github.com/mitchellh/go-homedir"
 )
 
-func setup(t *testing.T) (server *httptest.Server, dir string) {
-	tempDir, err := ioutil.TempDir("/var/tmp", "venus-auth")
-	if err != nil {
-		t.Fatal(err)
-	}
+func setup(t *testing.T) (server *httptest.Server, dir string, token string) {
+	tempDir := t.TempDir()
 	log.Infof("create storage temp dir: %s", tempDir)
 
-	cnf, err := config.DefaultConfig()
-	if err != nil {
-		t.Fatal(err)
-	}
-	cnf.DB.DSN = tempDir
-
-	dir, err = homedir.Expand(tempDir)
+	cnf := config.DefaultConfig()
+	dir, err := homedir.Expand(tempDir)
 	if err != nil {
 		t.Fatalf("could not expand repo location error:%s", err)
 	} else {
@@ -36,14 +27,18 @@ func setup(t *testing.T) (server *httptest.Server, dir string) {
 	gin.SetMode(gin.DebugMode)
 	dataPath := path.Join(dir, "data")
 
-	app, err := auth.NewOAuthApp(cnf.Secret, dataPath, cnf.DB)
+	app, err := auth.NewOAuthApp(dataPath, cnf.DB)
 	if err != nil {
-		t.Fatalf("Failed to init venus-auth: %s", err)
+		t.Fatalf("Failed to init sophon-auth: %s", err)
+	}
+	token, err = app.GetDefaultAdminToken()
+	if err != nil {
+		t.Fatalf("Failed to get default admin token: %s", err)
 	}
 
 	router := auth.InitRouter(app)
 	srv := httptest.NewServer(router)
-	return srv, tempDir
+	return srv, tempDir, token
 }
 
 func shutdown(t *testing.T, tempDir string) {
