@@ -2,11 +2,12 @@
 package integrate
 
 import (
+	"context"
 	"testing"
 
-	"github.com/filecoin-project/venus-auth/auth"
-	"github.com/filecoin-project/venus-auth/jwtclient"
-	"github.com/filecoin-project/venus-auth/storage"
+	"github.com/ipfs-force-community/sophon-auth/auth"
+	"github.com/ipfs-force-community/sophon-auth/jwtclient"
+	"github.com/ipfs-force-community/sophon-auth/storage"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -21,14 +22,14 @@ func TestRateLimitApis(t *testing.T) {
 }
 
 func setupAndAddRateLimits(t *testing.T) (*jwtclient.AuthClient, string) {
-	server, tmpDir := setup(t)
+	server, tmpDir, token := setup(t)
 
-	client, err := jwtclient.NewAuthClient(server.URL)
+	client, err := jwtclient.NewAuthClient(server.URL, token)
 	assert.Nil(t, err)
 
 	// Create a user
 	userName := "Rennbon"
-	_, err = client.CreateUser(&auth.CreateUserRequest{Name: userName})
+	_, err = client.CreateUser(context.TODO(), &auth.CreateUserRequest{Name: userName})
 	assert.Nil(t, err)
 
 	// Insert rate limit
@@ -43,7 +44,7 @@ func setupAndAddRateLimits(t *testing.T) (*jwtclient.AuthClient, string) {
 		},
 	}
 
-	upsertResp, err := client.UpsertUserRateLimit(&upsertReq)
+	upsertResp, err := client.UpsertUserRateLimit(context.TODO(), &upsertReq)
 	assert.Nil(t, err)
 	assert.Equal(t, upsertReq.Id, upsertResp)
 
@@ -54,7 +55,7 @@ func testUpsertUserRateLimit(t *testing.T) {
 	c, tmpDir := setupAndAddRateLimits(t)
 
 	// `ShouldBind` failed
-	_, err := c.UpsertUserRateLimit(&auth.UpsertUserRateLimitReq{})
+	_, err := c.UpsertUserRateLimit(context.TODO(), &auth.UpsertUserRateLimitReq{})
 	assert.Error(t, err)
 
 	shutdown(t, tmpDir)
@@ -67,13 +68,13 @@ func testGetRateLimit(t *testing.T) {
 	userName := "Rennbon"
 	reqId := "794fc9a4-2b80-4503-835a-7e8e27360b3d"
 	// Get user rate limit
-	getResp, err := client.GetUserRateLimit(userName, reqId)
+	getResp, err := client.GetUserRateLimit(context.Background(), userName, reqId)
 	assert.Nil(t, err)
 	assert.Equal(t, 1, len(getResp))
 	assert.Equal(t, reqId, getResp[0].Id)
 
 	// `ShouldBind` failed
-	_, err = client.GetUserRateLimit("", "")
+	_, err = client.GetUserRateLimit(context.Background(), "", "")
 	assert.Error(t, err)
 }
 
@@ -84,16 +85,16 @@ func testDeleteRateLimit(t *testing.T) {
 	userName := "Rennbon"
 	reqId := "794fc9a4-2b80-4503-835a-7e8e27360b3d"
 	// Delete rate limit
-	deleteResp, err := client.DelUserRateLimit(&auth.DelUserRateLimitReq{Name: userName, Id: reqId})
+	deleteResp, err := client.DelUserRateLimit(context.TODO(), &auth.DelUserRateLimitReq{Name: userName, Id: reqId})
 	assert.Nil(t, err)
 	assert.Equal(t, deleteResp, reqId)
 
 	// Try to get deleted rate limit
-	getResp, err := client.GetUserRateLimit(userName, reqId)
+	getResp, err := client.GetUserRateLimit(context.Background(), userName, reqId)
 	assert.Nil(t, err)
 	assert.Equal(t, 0, len(getResp))
 
 	// if there is an error deleting user rate limits
-	_, err = client.DelUserRateLimit(&auth.DelUserRateLimitReq{})
+	_, err = client.DelUserRateLimit(context.TODO(), &auth.DelUserRateLimitReq{})
 	assert.Error(t, err)
 }
