@@ -3,7 +3,7 @@ package log
 import (
 	"os"
 
-	"github.com/filecoin-project/venus-auth/config"
+	"github.com/ipfs-force-community/sophon-auth/config"
 	"github.com/sirupsen/logrus"
 )
 
