@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/filecoin-project/venus-auth/config"
+	"github.com/ipfs-force-community/sophon-auth/config"
 	"github.com/sirupsen/logrus"
 )
 
